Create The Houdini Code Used to Generate the Juliabulb
Unlike MATLAB, the code in Houdini is not enough to generate the figures of the Mandelbulb, Juliabulb, or Logisticbulb. Please watch a YouTube video to do other set-ups (https://www.youtube.com/watch?v=DLjZXYi6smc). The power $n$ and the iteration of $v\mapsto v^n + c$ can be changed in the CONTROLLER.